+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Renamed `LintMeta` to `Lint` and made it exhaustive and user
   constructible
+  - Added `message` member to it
 
 
 0.1.3